Spells and health related
Reordered spells folders and files.
Added on trigger enter for continuous spells. Moved those spells to is trigger.
Replaced the flamethrower sounds.
Reworked the player ui controller for health then moved it to be on the HUD object.
Created a new player health component. It does not currently derive from the player stats SO.
Updated damage in projectile mover for player to use the new player health.
Changed some of the prefab references for the new health setup.
Added event listener for health plus the invoke logic.