main fixes
fixed everything but the cursor position object issue. idk how you drive the position of the actual gameobject but that's the issue is it's locationally mislocated. I will flip directions once you fix that, because right now it's hard to know what's wrong when the object is itself mislocated wrong.

I removed all if stuff besides isreadytocast on playercontroller for when to cast. i fixed the method call names on audio and uncommented them.

I updated the player prefab to have the missing ui stuff for spells. i also added the spawn points and new references. tweaked the tree scripts to fix the ui bug where skillpoints were not updating in one scenario.

i updated the spell listener prefab to remove the event we no longer need for cast was invoked or whatever.

i added lighting controller to the game manager. please clean up the prefab mess. save your changes or whatever. it's becoming cumbersome.

abysall fang was for some reason disabled and missing the icon and reticle references. i re-added those and made it unlocked.