Main fix 2
modified tc spell to have a default spread value that is passed to the spread value. this is part of the massive workaround on most spells to cope with not having a save system to save values and SO have persistent value changes. just backend polish.

reset bolt spread value to 9 after fixing.